api/requirements.txt
fastapi>=0.115
uvicorn>=0.34
sqlalchemy>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
